+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- Comment widget in pages mode without field-list has hidden comment field and shows empty page.
+
 [4.16.1] - 2016-12-05
 ---------------------
 ##### Fixed